+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Remove stale transpiled files when running `ruby-next nextify`. ([@palkan][])
+
 - Add Ruby 2.4 support. ([@palkan][])
 
 APIs for <2.5 must be backported via [backports][] gem. Refinements are not supported.